backend/requirements.txt
fastapi==0.115.0
pytest==8.3.5
httpx==0.28.1
python-multipart==0.0.9
uvicorn[standard]==0.30.6
torch==2.4.1
torchvision==0.19.1
pillow==10.4.0
pillow-heif==0.21.0
pandas==2.2.2
numpy==1.26.4
openai==1.58.1
